Implemented VacuumsAmount Apearance

+++ Assets/Scripts/UI/UpgradesShop/Upgraders/VacuumsAmount/VacuumsAmountUpgradeStock.cs
@@ -3,5 +3,5 @@
 [CreateAssetMenu(fileName = "VacuumsAmountUpgradeStock", menuName = "Upgrades/Stocks/VacuumsAmountUpgradeStock")]
 public class VacuumsAmountUpgradeStock : UpgradeStockBase
 {
-    public int VacuumsAmount;
+    public float VacuumsAmount;
 }
